测试前临时备份

+++ TX-8600/src/user/conn_process.c
@@ -492,10 +492,12 @@ void conn_overtime_close(){
         //-------------------------------------------------------
         //连接超时，删除节点，关闭连接
         if(conn_list_tmp->over_time>CONN_OVERTIME){
-            xtcp_debug_printf("conn timeout %x\n",conn_list_tmp->conn.id);
-            user_xtcp_close(conn_list_tmp->conn);
-            mes_list_close(conn_list_tmp->conn.id);
-            delete_conn_node(conn_list_tmp->conn.stack_conn);
+            if((conn_list_tmp->conn.id != g_sys_val.debug_conn.id) || (g_sys_val.eth_debug_f==0)){  
+                xtcp_debug_printf("conn timeout %x\n",conn_list_tmp->conn.id);
+                user_xtcp_close(conn_list_tmp->conn);
+                mes_list_close(conn_list_tmp->conn.id);
+                delete_conn_node(conn_list_tmp->conn.stack_conn);
+            }
         }    
         //-------------------------------------------------------
         conn_list_tmp = conn_next_p;
